fastapi/requirements.txt
torch
torchvision
Pillow
gdown
fastapi